Enhance Lua 5.1 core (stopgap until LuaJIT)
Enhances the copy of Lua 5.1.5 with small additions, performance improvements, and bugfixes from various places with the goal of being a stopgap until a modified fork of LuaJIT in integrated.
things of note are:
- Compound operators (`+=`,`-=`,`*=`, `/=`, `%=`, `^=`).
- Continue statement.
- Fully resumeable coroutines.
- Bitwise operations module/library.
- Extra math constants (`math.tau` and `math.e`).
- Small memory reduction and performance boost from NAN packing.
- `math.log` now accepts an optional base number.
- Support binary number literals.
- Add warning/error for duplicate variable declaration.
- Remove old vararg style compat (This is how it's like in LuaJIT).
- Tables now honor `__len`.
- New `rawlen()` function.
- String matching pattern %g added.

below is a dump of various URLs from which the changes in this commit are from or based on (NOTE: some of these don't exactly match up with this commit due to changes I made after the fact for various reasons such as compat with LuaJIT or otherwise):
https://bitop.luajit.org/
https://coco.luajit.org/
https://lua-users.org/wiki/YieldableForLoops
https://github.com/lua/lua/commit/5019b2dd205751c6f29c4ced2b6f1e22c096fb83
https://github.com/lua/lua/commit/897573983977e2bdfe405bd056abccc1db1e0f8d
https://github.com/lua/lua/commit/f292760f12022a83cf01e788482a264aeeb3c276
https://github.com/lua/lua/commit/5a7dee0c5d8dfa9a4d95c216f7c38d9f2b1cd441
https://github.com/lua/lua/commit/3c710f056b683136a342bc5bf88d852717cd0d33
https://github.com/dubiousjim/luafiveq/blob/master/patches/os-exit.patch
https://github.com/dubiousjim/luafiveq/blob/master/patches/table-len.patch
https://github.com/dubiousjim/luafiveq/blob/master/patches/yieldable-for-loops.patch
http://lua-users.org/files/wiki_insecure/power_patches/5.1/jh-lua-binoctal-5.1.4.patch
http://lua-users.org/files/wiki_insecure/power_patches/5.1/pack_value.patch
http://lua-users.org/wiki/ContinueProposal
http://www.corsix.org/luacontinue/continue.patch
http://lua-users.org/lists/lua-l//2011-06/msg01223.html
https://github.com/idmillington/lua/commit/0a817b9e0b1f37e81db0fd615aeddcb90e988629

Something to note is that Coco allows more things to yield then modern LuaJIT does so watch out with compat related to that in scripts. A combination of `coroutine.coco` from Coco and `jit.version` from LuaJIT can be used to detect which coroutine model is in use (aka how restrictive yielding is) so that code can be made compatible with both.
Another thing to note is the Lua 5.1 continue patch used here is more restrictive then the LuaJIT continue patch.